Fixed schema caching.
Previously, a non-persistent counter would be used as part of the
cache key, and changes made to this counter in one process would
obviously not propagate to other processes.

Instead, the cache key now includes the schema and subtypes which are
both retrieved from a FTI-specific volatile cache that uses the
modification time as its cache key.

Note that the tests are failing in this particular changeset. It's a
little tedious to instruct the mocker to respond naturally to the
volatile caching attempts. Maybe the passthrough-mode can be used,
combined with a single test case that examines the cache invariant.