utils.createContent honors behaviors
createContent was just doing set attribute on all key passed.
This works because by default we don't have complex behavior.
I'm working with plone.app.event on a project and creating
an event with createContent doesn't works, it broke the timezone.

Behaviors has the responsabilities to handle their data.
Here start,end,timezone must be handled by IEventBasic behavior
or you will not have a consistent data. The bug we have is the
date is translated of 1 ours (UTC+1).

Because createContent is not tested at all in plone.dexterity
and only use a bit in plone.app.dexterity I submit it as
a pull request for review.
Thanks to review it and let me know if you have any problem with
this change.