fix: avoid setting a default value on methods
If a Schema Interface has a method in it, i.e. to be used as a
constraint for another field, etc. the  function
would trip over it while trying to get a default value for it.

picked from https://github.com/plone/plone.dexterity/commit/80735b5563bed96f67557b328eb31db6acf6c283