Fixes schema name generated in Python 2
As of plone.dexterity 2.9.10, _p_mtime is used in the schema name. To
transform the _p_mtime into a string, the str function was used.
However, in Python 2, the str function rounds floats when it's applied
to large floats:

Python 2:
>>> str(1637689348.9999528)
'1637689349.0'

Python 3:
>>> str(1637689348.9999528)
'1637689348.9999528'

This was causing the schema names in Python 2 to take an unexpected
format, causing errors.
So, we need to use the repr function, which doesn't round floats.